requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
timm>=0.9.0

# Model Optimization
bitsandbytes>=0.41.0
accelerate>=0.24.0

# Computer Vision
opencv-python>=4.8.0
ultralytics>=8.0.0
decord>=0.6.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0

# Configuration
pyyaml>=6.0.0

# Utilities
tqdm>=4.65.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
